Move simd functions to qubitvector.hpp
