+++ CMakeLists.txt
@@ -206,7 +206,7 @@ endif()
 
 if(APPLE)
 	enable_cxx_compiler_flag_if_supported("-mavx")
-	set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -DSIMD")
+	set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -DSIMD_MAC")
 endif()
 
 # Tests
